+++ test/EliteChroma.Core.Tests/KeyMappings.Test.cs
@@ -65,10 +65,11 @@ namespace EliteChroma.Core.Tests
         [Theory]
         [InlineData("Key_Escape", "en-US", true, VirtualKey.VK_ESCAPE)]
         [InlineData("Key_ß", "de-DE", true, VirtualKey.VK_OEM_4)]
+        [InlineData("Key_º", "es-ES", true, VirtualKey.VK_OEM_5)]
         [InlineData("Key_INVALID_KEY_NAME", "en-US", false, (VirtualKey)0)]
         public void TryGetVirtualKeyReturnsExpectedValues(string keyName, string keyboardLayout, bool expectedOk, Enum expectedKey)
         {
-            var ok = Elite.Internal.KeyMappings.TryGetKey(keyName, keyboardLayout, out var virtualKey, NativeMethodsKeyboardMock.Instance);
+            var ok = Elite.Internal.KeyMappings.TryGetKey(keyName, keyboardLayout, false, out var virtualKey, NativeMethodsKeyboardMock.Instance);
 
             Assert.Equal(expectedOk, ok);
             Assert.Equal((VirtualKey)expectedKey, virtualKey);
@@ -78,10 +79,31 @@ namespace EliteChroma.Core.Tests
         [InlineData("Key_Escape", "en-US", true, Key.Escape)]
         [InlineData("Key_ß", "de-DE", true, Key.OemMinus)]
         [InlineData("Key_¤", "en-US", false, (Key)0)]
+        [InlineData("Key_º", "es-ES", true, Key.OemTilde)]
         [InlineData("Key_INVALID_KEY_NAME", "en-US", false, (Key)0)]
         public void TryGetChromaKeyReturnsExpectedValues(string keyName, string keyboardLayout, bool expectedOk, Key expectedKey)
         {
-            var ok = Core.Internal.KeyMappings.TryGetKey(keyName, keyboardLayout, out var chromaKey, NativeMethodsKeyboardMock.Instance);
+            var ok = Core.Internal.KeyMappings.TryGetKey(keyName, keyboardLayout, false, out var chromaKey, NativeMethodsKeyboardMock.Instance);
+
+            Assert.Equal(expectedOk, ok);
+            Assert.Equal(expectedKey, chromaKey);
+        }
+
+        [Theory]
+        [InlineData("Key_Grave", "es-ES", true, VirtualKey.VK_OEM_5)]
+        public void TryGetVirtualKeyReturnsEnUSOverrides(string keyName, string keyboardLayout, bool expectedOk, Enum expectedKey)
+        {
+            var ok = Elite.Internal.KeyMappings.TryGetKey(keyName, keyboardLayout, true, out var virtualKey, NativeMethodsKeyboardMock.Instance);
+
+            Assert.Equal(expectedOk, ok);
+            Assert.Equal((VirtualKey)expectedKey, virtualKey);
+        }
+
+        [Theory]
+        [InlineData("Key_Grave", "es-ES", true, Key.OemTilde)]
+        public void TryGetChromaKeyReturnsEnUSOverrides(string keyName, string keyboardLayout, bool expectedOk, Key expectedKey)
+        {
+            var ok = Core.Internal.KeyMappings.TryGetKey(keyName, keyboardLayout, true, out var chromaKey, NativeMethodsKeyboardMock.Instance);
 
             Assert.Equal(expectedOk, ok);
             Assert.Equal(expectedKey, chromaKey);